requirements.txt
requests
lxml
pyyaml
beautifulsoup4
pillow
schedule
pymupdf
jinja2
pylatexenc
